+++ adhoc-core/src/main/java/adhoc/objective/ObjectiveDto.java
@@ -57,11 +57,11 @@ public class ObjectiveDto {
     private String name;
 
     @NotNull
-    private Float x;
+    private Double x;
     @NotNull
-    private Float y;
+    private Double y;
     @NotNull
-    private Float z;
+    private Double z;
 
     @Min(1)
     private Long initialFactionId;
